Consolidating Dashboard header test files
