+++ frontend/src/metabase/dashboard/containers/DashboardHeader.unit.spec.tsx
@@ -1,69 +1,138 @@
-import fetchMock from "fetch-mock";
+import userEvent from "@testing-library/user-event";
+import { getDefaultTab } from "metabase/dashboard/actions";
 import {
   renderWithProviders,
   screen,
   waitForElementToBeRemoved,
+  within,
 } from "__support__/ui";
+
+import {
+  setupBookmarksEndpoints,
+  setupCollectionByIdEndpoint,
+  setupCollectionsEndpoints,
+} from "__support__/server-mocks";
 import {
   createMockCollection,
   createMockDashboard,
 } from "metabase-types/api/mocks";
+import { Dashboard } from "metabase-types/api";
+import { setupEnterpriseTest } from "__support__/enterprise";
 import DashboardHeader from "./DashboardHeader";
 
-const setup = () => {
-  fetchMock.get(
-    "path:/api/collection/10",
-    createMockCollection({
-      name: "Custom Reports",
-      id: 10,
-      type: "instance-analytics",
+const TEST_DASHBOARD = createMockDashboard();
+
+const TEST_DASHBOARD_WITH_TABS = createMockDashboard({
+  id: 2,
+  ordered_tabs: [
+    getDefaultTab({ tabId: 1, dashId: 1, name: "Tab 1" }),
+    getDefaultTab({
+      tabId: 2,
+      dashId: 1,
+      name: "Tab 2",
     }),
-  );
+  ],
+});
 
-  fetchMock.get("path:/api/bookmark", []);
+const ISNTANCE_ANALYTICS_DASHBOARD = createMockDashboard({
+  id: 3,
+  collection_id: 10,
+});
 
-  const dashboard = createMockDashboard({
-    collection_id: 10,
-  });
+const INSTANCE_ANALYTICS_COLLECTION = createMockCollection({
+  name: "Custom Reports",
+  id: 10,
+  type: "instance-analytics",
+});
 
-  renderWithProviders(
-    <DashboardHeader
-      dashboard={dashboard}
-      saveDashboardAndCards={jest.fn()}
-      setDashboardAttribute={jest.fn()}
-      onEditingChange={jest.fn()}
-      onRefreshPeriodChange={jest.fn()}
-      onNightModeChange={jest.fn()}
-      onFullscreenChange={jest.fn()}
-      onSharingClick={jest.fn()}
-      isFullscreen={false}
-      isNightMode={false}
-      setRefreshElapsedHook={jest.fn()}
-      addCardToDashboard={jest.fn()}
-      addTextDashCardToDashboard={jest.fn()}
-      addLinkDashCardToDashboard={jest.fn()}
-      isEditable={false}
-      isEditing={false}
-      databases={[]}
-      location={{ pathname: "/dashboard/2" }}
-    />,
-  );
+const setup = async ({ dashboard }: { dashboard: Dashboard }) => {
+  setupCollectionsEndpoints({collections: []});
+  setupCollectionByIdEndpoint({ collections: [INSTANCE_ANALYTICS_COLLECTION] });
+  setupBookmarksEndpoints([]);
+
+  const dashboardHeaderProps = {
+    dashboard,
+    isEditable: true,
+    isEditing: false,
+    isFullscreen: false,
+    isNavBarOpen: false,
+    isNightMode: false,
+    isAdditionalInfoVisible: false,
+    refreshPeriod: 0,
+    setRefreshElapsedHook: jest.fn(),
+    addCardToDashboard: jest.fn(),
+    addTextDashCardToDashboard: jest.fn(),
+    addLinkDashCardToDashboard: jest.fn(),
+    fetchDashboard: jest.fn(),
+    saveDashboardAndCards: jest.fn(),
+    setDashboardAttribute: jest.fn(),
+    onEditingChange: jest.fn(),
+    onRefreshPeriodChange: jest.fn(),
+    onNightModeChange: jest.fn(),
+    onFullscreenChange: jest.fn(),
+    onSharingClick: jest.fn(),
+    onChangeLocation: jest.fn(),
+    toggleSidebar: jest.fn(),
+    sidebar: {
+      name: "",
+      props: {},
+    },
+    location: {},
+    setSidebar: jest.fn(),
+    closeSidebar: jest.fn(),
+    addActionToDashboard: jest.fn(),
+    databases: {},
+  };
+
+  renderWithProviders(<DashboardHeader {...dashboardHeaderProps} />);
+
+  await waitForElementToBeRemoved(() => screen.queryByText("Loading..."));
 };
 
 describe("dashboard header", () => {
-  it("should render the correct buttons", async () => {
-    setup();
-    await waitForElementToBeRemoved(() => screen.queryByText("Loading..."));
-    expect(screen.getByText("Make a copy")).toBeInTheDocument();
-    expect(screen.getByRole("img", { name: /beaker/i })).toBeInTheDocument();
+  it("should display `Export as PDF` when there is a single dashboard tab", async () => {
+    await setup({
+      dashboard: TEST_DASHBOARD,
+    });
 
-    //Other buttons
-    expect(
-      screen.getByRole("button", { name: /bookmark/i }),
-    ).toBeInTheDocument();
-    expect(screen.getByRole("button", { name: /info/i })).toBeInTheDocument();
-    expect(
-      screen.getByRole("button", { name: /fullscreen/i }),
-    ).toBeInTheDocument();
+    userEvent.click(screen.getByLabelText("dashboard-menu-button"));
+
+    const exportPdfButton = within(
+      screen.getByTestId("dashboard-export-pdf-button"),
+    );
+    expect(exportPdfButton.getByText("Export as PDF")).toBeInTheDocument();
+  });
+
+  it("should display `Export tab as PDF` when there are multiple dashboard tabs", async () => {
+    await setup({
+      dashboard: TEST_DASHBOARD_WITH_TABS,
+    });
+
+    userEvent.click(screen.getByLabelText("dashboard-menu-button"));
+
+    const exportPdfButton = within(
+      screen.getByTestId("dashboard-export-pdf-button"),
+    );
+    expect(exportPdfButton.getByText("Export tab as PDF")).toBeInTheDocument();
+  });
+
+  describe("EE", () => {
+    it("should render the correct buttons for instance analytics dashboard", async () => {
+      setupEnterpriseTest();
+      await setup({
+        dashboard: ISNTANCE_ANALYTICS_DASHBOARD,
+      });
+      expect(screen.getByText("Make a copy")).toBeInTheDocument();
+      expect(screen.getByRole("img", { name: /beaker/i })).toBeInTheDocument();
+
+      //Other buttons
+      expect(
+        screen.getByRole("button", { name: /bookmark/i }),
+      ).toBeInTheDocument();
+      expect(screen.getByRole("button", { name: /info/i })).toBeInTheDocument();
+      expect(
+        screen.getByRole("button", { name: /fullscreen/i }),
+      ).toBeInTheDocument();
+    });
   });
 });
